Add support for vertical scrolling across nametables
This expands the nametable/attribute tables to be arrays with each level
being a different table number. The GL program draws three quads
vertically in order to handle scrolling past nametable B.

( 1, 1) +-------------+ scroll=0
        |             |
        |             |
        |      A      |
        |             |
        |             |
(-1,-1) +-------------+ scroll=240
        |             |
        |             |
        |      B      |
        |             |
        |             |
(-3,-3) +-------------+ scroll=0
        |             |
        |             |
        |      A      |
        |             |
        |             |
(-5,-5) +-------------+ scroll=240

+++ src/ui/opengl/window.h
@@ -23,7 +23,7 @@ class OpenGLWindow : public Window {
   // Window implementation:
   bool IsClosed() override { return is_closed_; }
   void SetTitle(std::string title) override;
-  void SetScroll(uint8_t x, uint8_t y) override;
+  void SetScroll(int x, int y) override;
   void SetNametable(int num, std::vector<uint8_t>& nametable) override;
   void SetAttributeTable(int num, std::vector<uint8_t>& table) override;
   void SetFramePalette(std::vector<uint8_t>& palette) override;
