+++ src/ui/window.h
@@ -10,7 +10,7 @@ class Window {
   virtual ~Window() {}
   virtual bool IsClosed() = 0;
   virtual void SetTitle(std::string title) = 0;
-  virtual void SetScroll(uint8_t x, uint8_t y) = 0;
+  virtual void SetScroll(int x, int y) = 0;
   virtual void SetNametable(int num, std::vector<uint8_t>& nametable) = 0;
   virtual void SetAttributeTable(int num, std::vector<uint8_t>& table) = 0;
   virtual void SetFramePalette(std::vector<uint8_t>& palette) = 0;
@@ -23,7 +23,7 @@ class FakeWindow : public Window {
   // Window implementation:
   bool IsClosed() override { return false; }
   void SetTitle(std::string title) override {}
-  void SetScroll(uint8_t x, uint8_t y) override {}
+  void SetScroll(int x, int y) override {}
   void SetNametable(int num, std::vector<uint8_t>& nametable) override {}
   void SetAttributeTable(int num, std::vector<uint8_t>& table) override {}
   void SetFramePalette(std::vector<uint8_t>& palette) override {}
